third_party/blink/web_tests/http/tests/images/document-policy-unoptimized-lossless-images-expected.txt

CONSOLE ERROR: Document policy violation: unoptimized-lossless-images is not allowed in this document.
